// src/Moryx.Workplans.Web/app/src/app/api/models/workplan-session-model.ts
/*
 * Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 * Licensed under the Apache License, Version 2.0
*/

/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { WorkplanNodeModel } from '../models/workplan-node-model';
import { WorkplanState } from '../models/workplan-state';
export interface WorkplanSessionModel {
  name?: string | null;
  nodes?: Array<WorkplanNodeModel> | null;
  sessionToken?: string | null;
  state?: WorkplanState;
  version?: number;
  workplanId?: number;
}